[ApiBundle] Move LocaleIsUsedException into ApiBundle
